Refactor ListBoxItemChipList into own file
